chore: move out renderer typings from styles package
