+++ packages/fluentui/docs/src/components/VariableResolver/useEnhancedRenderer.ts
@@ -1,5 +1,6 @@
 import { callable } from '@fluentui/styles';
-import { ComponentSlotStylesPrepared, Renderer } from '@fluentui/react-northstar';
+import { Renderer } from '@fluentui/react-northstar-styles-renderer';
+import { ComponentSlotStylesPrepared } from '@fluentui/react-northstar';
 import flat from 'flat';
 import * as _ from 'lodash';
 import * as React from 'react';
