+++ packages/fluentui/react-bindings/package.json
@@ -12,6 +12,7 @@
     "@fluentui/react-component-ref": "^0.49.0",
     "@fluentui/react-compose": "^0.11.1",
     "@fluentui/react-northstar-fela-renderer": "^0.49.0",
+    "@fluentui/react-northstar-styles-renderer": "^0.49.0",
     "@fluentui/state": "^0.49.0",
     "@fluentui/styles": "^0.49.0",
     "@uifabric/utilities": "^7.21.0",
