+++ packages/fluentui/react-bindings/src/styles/types.ts
@@ -1,11 +1,6 @@
-import {
-  DebugData,
-  ICSSInJSStyle,
-  PropsWithVarsAndStyles,
-  ThemeInput,
-  ThemePrepared,
-  Renderer,
-} from '@fluentui/styles';
+import { Renderer } from '@fluentui/react-northstar-styles-renderer';
+import { DebugData, ICSSInJSStyle, PropsWithVarsAndStyles, ThemeInput, ThemePrepared } from '@fluentui/styles';
+
 import { Telemetry } from '../telemetry/types';
 
 // Notice:
