+++ packages/fluentui/react-northstar-fela-renderer/package.json
@@ -5,7 +5,7 @@
   "bugs": "https://github.com/microsoft/fluentui/issues",
   "dependencies": {
     "@babel/runtime": "^7.7.6",
-    "@fluentui/styles": "^0.49.0",
+    "@fluentui/react-northstar-styles-renderer": "^0.49.0",
     "css-in-js-utils": "^3.0.0",
     "fela": "^10.6.1",
     "fela-plugin-embedded": "^10.6.1",
