+++ packages/fluentui/react-northstar-fela-renderer/src/felaStylisEnhancer.ts
@@ -2,7 +2,7 @@ import { RULE_TYPE } from 'fela-utils';
 // @ts-ignore
 import _Stylis from 'stylis';
 
-import { Renderer, RendererChange } from './types';
+import { FelaRenderer, FelaRendererChange } from './types';
 
 // `stylis@3` is a CJS library, there are known issues with them:
 // https://github.com/rollup/rollup/issues/1267#issuecomment-446681320
@@ -18,10 +18,10 @@ const stylis = new Stylis({
   semicolon: false,
 });
 
-const felaStylisEnhancer = (renderer: Renderer) => {
+const felaStylisEnhancer = (renderer: FelaRenderer) => {
   const existingEmitChange = renderer._emitChange.bind(renderer);
 
-  renderer._emitChange = (change: RendererChange) => {
+  renderer._emitChange = (change: FelaRendererChange) => {
     if (change.type === RULE_TYPE) {
       const prefixed: string = stylis('', change.declaration);
