+++ packages/fluentui/react-northstar-fela-renderer/src/types.ts
@@ -1,14 +1,15 @@
 import { IRenderer } from 'fela';
+import { RendererParam } from '@fluentui/react-northstar-styles-renderer';
 import { ICSSInJSStyle } from '@fluentui/styles';
 
-export type Renderer = IRenderer & {
-  cache: Record<string, RendererChange>;
+export type FelaRenderer = IRenderer & {
+  cache: Record<string, FelaRendererChange>;
   selectorPrefix?: string;
 
   getNextRuleIdentifier: () => string;
   filterClassName: () => boolean;
 
-  _emitChange: (change: RendererChange) => void;
+  _emitChange: (change: FelaRendererChange) => void;
   _renderStyleToClassNames(
     { _className, ...style }: ICSSInJSStyle & { _className: string },
     pseudo?: string,
@@ -17,7 +18,7 @@ export type Renderer = IRenderer & {
   ): string;
 };
 
-export type RendererChange = {
+export type FelaRendererChange = {
   type: 'RULE' | 'KEYFRAME' | 'FONT' | 'STATIC' | 'CLEAR';
   className: string;
   selector: string;
@@ -26,3 +27,5 @@ export type RendererChange = {
   media: string;
   support: string;
 };
+
+export type FelaRendererParam = Omit<RendererParam, 'direction'> & { theme: { direction: RendererParam['direction'] } };
