+++ packages/fluentui/react-northstar-fela-renderer/test/felaDisableAnimationsPlugin-test.ts
@@ -1,5 +1,5 @@
-import felaDisableAnimationsPlugin from '../src/felaDisableAnimationsPlugin';
 import { ICSSInJSStyle } from '@fluentui/styles';
+import felaDisableAnimationsPlugin from '../src/felaDisableAnimationsPlugin';
 
 const stylesWithAnimationShorthand: ICSSInJSStyle = {
   animation: 'k1',
