+++ packages/fluentui/react-northstar-styles-renderer/.eslintignore
@@ -1,0 +1,4 @@
+coverage/
+dist/
+lib/
+node_modules/
