+++ packages/fluentui/react-northstar-styles-renderer/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../../scripts/eslint/index"],
+  "root": true
+}
