+++ packages/fluentui/react-northstar-styles-renderer/.gulp.js
@@ -1,0 +1,1 @@
+module.exports = require('@uifabric/build/gulp/.gulp');
