+++ packages/fluentui/react-northstar-styles-renderer/babel.config.js
@@ -1,0 +1,1 @@
+module.exports = api => require('@uifabric/build/babel')(api);
