+++ packages/fluentui/react-northstar-styles-renderer/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@fluentui/react-northstar-styles-renderer",
+  "description": "A noop renderer for FluentUI React Northstar with typings.",
+  "version": "0.49.0",
+  "bugs": "https://github.com/microsoft/fluentui/issues",
+  "dependencies": {
+    "@babel/runtime": "^7.7.6",
+    "@fluentui/styles": "^0.49.0"
+  },
+  "devDependencies": {
+    "@types/react": "16.8.25",
+    "@uifabric/build": "^7.0.0",
+    "lerna-alias": "^3.0.3-0",
+    "react": "16.8.6"
+  },
+  "files": [
+    "dist"
+  ],
+  "homepage": "https://github.com/microsoft/fluentui/tree/master/packages/fluentui/react-northstar-styles-renderer",
+  "jsnext:main": "dist/es/index.js",
+  "license": "MIT",
+  "main": "dist/commonjs/index.js",
+  "module": "dist/es/index.js",
+  "peerDependencies": {
+    "react": "^16.8.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": "microsoft/fluentui.git",
+  "scripts": {
+    "build": "gulp bundle:package:no-umd",
+    "clean": "gulp bundle:package:clean",
+    "lint": "eslint --ext .js,.ts,.tsx .",
+    "lint:fix": "yarn lint --fix"
+  },
+  "sideEffects": false,
+  "types": "dist/es/index.d.ts"
+}
