+++ packages/fluentui/react-northstar-styles-renderer/src/index.ts
@@ -1,0 +1,2 @@
+export { noopRenderer } from './noopRenderer';
+export * from './types';
