+++ packages/fluentui/react-northstar/package.json
@@ -15,6 +15,7 @@
     "@fluentui/react-compose": "^0.11.1",
     "@fluentui/react-context-selector": "^0.49.0",
     "@fluentui/react-icons-northstar": "^0.49.0",
+    "@fluentui/react-northstar-styles-renderer": "^0.49.0",
     "@fluentui/react-proptypes": "^0.49.0",
     "@fluentui/state": "^0.49.0",
     "@fluentui/styles": "^0.49.0",
