+++ packages/fluentui/react-northstar/tsconfig.json
@@ -33,6 +33,9 @@
       "path": "../react-context-selector"
     },
     {
+      "path": "../react-northstar-styles-renderer"
+    },
+    {
       "path": "../react-proptypes"
     }
   ]
